+++ pp/components/coupler_full.py
@@ -6,7 +6,6 @@ import picwriter.components as pc
 import pp
 from pp.component import Component
 from pp.components.waveguide_template import strip
-from pp.picwriter_to_component import picwriter_to_component
 from pp.types import ComponentFactory
 
 
@@ -80,7 +79,7 @@ def coupler_full(
         direction=direction,
     )
 
-    return picwriter_to_component(c)
+    return pp.component_from.picwriter(c)
 
 
 if __name__ == "__main__":