+++ pp/component_from/__init__.py
@@ -1,4 +1,6 @@
+from pp.component_from.gds import gds
+from pp.component_from.netlist import netlist
 from pp.component_from.phidl import phidl
 from pp.component_from.picwriter import picwriter
 
-__all__ = ["phidl", "picwriter"]
+__all__ = ["phidl", "picwriter", "gds", "netlist"]