+++ pp/components/mzi1x2.py
@@ -6,7 +6,6 @@ from pp.components.mmi1x2 import mmi1x2
 from pp.components.mzi2x2 import mzi_arm
 from pp.components.straight import straight as straight_function
 from pp.components.straight_heater import straight_with_heater
-from pp.netlist_to_gds import netlist_to_component
 from pp.port import select_electrical_ports
 from pp.routing.route_ports_to_side import route_ports_to_side
 from pp.types import ComponentFactory
@@ -105,7 +104,7 @@ def mzi1x2(
             "E_BOT_3": ("arm_bot", "E_3"),
         }
 
-        component = netlist_to_component(components, connections, ports_map)
+        component = pp.component_from.netlist(components, connections, ports_map)
         # Need to connect common ground and redefine electrical ports
 
         ports = component.ports
@@ -154,7 +153,7 @@ def mzi1x2(
 
     else:
         ports_map = {"W0": ("CP1", "W0"), "E0": ("CP2", "W0")}
-        component = netlist_to_component(components, connections, ports_map)
+        component = pp.component_from.netlist(components, connections, ports_map)
 
     return component