+++ CHANGELOG.md
@@ -43,6 +43,7 @@ from contracts import contract
 - move set_plot_options to pp.klive.set_plot_options, stop overriding phidl's set_plot_options in `pp.__init__`
 - move pp.merge_cells into pp.component_from.gdspaths and pp.component_from.gdsdir
 - waveguide accepts dict(component='fabc_nitride_cband')
+- add pp.remove and pp.read
 
 ## 2.6.5