+++ CHANGELOG.md
@@ -44,6 +44,7 @@ from contracts import contract
 - move pp.merge_cells into pp.component_from.gdspaths and pp.component_from.gdsdir
 - waveguide accepts dict(component='fabc_nitride_cband')
 - add pp.remove and pp.read
+- remove pp.gds
 
 ## 2.6.5